Minor refactoring
- clearly distinguish parcels machinery from VirtualFleet helpers
