+++ virtualargofleet/__init__.py
@@ -1,4 +1,5 @@
-from .virtualargofleet import velocityfield, virtualfleet
+from .virtualargofleet import VelocityField as velocityfield
+from .virtualargofleet import VirtualFleet as virtualfleet
 
 #
 __all__ = (
